:art: :: 서브 클래스를 통한 가독성 증진
