+++ Projects/Features/SearchFeature/Sources/CompositionalLayout/Enum/Section.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+internal enum Section: Int {
+    case youtube
+    case recommend
+    case popularList
+}
